Ensure that asynchronous exeuction message includes header(s) from the original request. Do not close database connection handled by GDAL from the OGC C base-vect-ops service.

+++ zoo-project/zoo-kernel/service.h
@@ -474,7 +474,7 @@ extern "C" {
   ZOO_DLL_EXPORT void addToMap(map*,const char*,const char*);
   ZOO_DLL_EXPORT void addToMapA(map*,const char*,const char*);
   ZOO_DLL_EXPORT void addIntToMap(map*,const char*,const int);
-  ZOO_DLL_EXPORT void addIntToMapArray(map*,const char*,int,const int);
+  ZOO_DLL_EXPORT void addIntToMapArray(map*,const char*,int,const unsigned long);
   ZOO_DLL_EXPORT map* addToMapWithSize(map*,const char*,const char*,int);
   ZOO_DLL_EXPORT void addMapToMap(map**,map*);
   ZOO_DLL_EXPORT void addMapToIoType(iotype**,map*);
